// morgenster-hospital-management-system/types.ts
export interface Patient {
    id?: string;
    hospitalNumber: string;
    name: string;
    surname: string;
    nationalId?: string; // Added National ID
    nationality: string; // Added Nationality
    passportNumber?: string; // Added Passport Number
    dateOfBirth: string;
    age: number;
    maritalStatus: string;
    gender: 'Male' | 'Female' | 'Other';
    countryOfBirth: string;
    phoneNumber: string;
    residentialAddress: string;
    nokName: string;
    nokSurname: string;
    nokPhoneNumber: string;
    nokAddress: string;
    registeredBy: string; // User ID
    registrationDate: string; // ISO string
    status: 'Admitted' | 'PendingDischarge' | 'Discharged';
    dischargeRequesterId?: string;
    financials: {
        totalBill: number;
        amountPaid: number;
        balance: number;
    };
    currentWardId?: string;
    currentWardName?: string;
    currentBedNumber?: number;
}